+++ lib/zaikio-client-helpers.rb
@@ -3,6 +3,4 @@
 require_relative "zaikio/client/helpers/version"
 
 require_relative "zaikio/error"
-require_relative "zaikio/client/model"
-require_relative "zaikio/client/helpers/json_parser"
-require_relative "zaikio/client/helpers/pagination"
+require_relative "zaikio/client"