+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+- Add `Zaikio::Client::Helpers::AuthorizationMiddleware` and `Zaikio::Client.with_token` to pass down bearer token to multiple clients (e.g. hub + procurement) at the same time
+- Add `Zaikio::Client::Helpers::Configuration` as an abstract configuration class
+- Add `Zaikio::Client.create_connection`
+
 ## [0.2.4] - 2022-03-29
 
 - Add support for Faraday 2.x